Feature/update user profile screen refactor

+++ app/src/main/java/com/mashup/zuzu/data/request/UpdateUsersRequest.kt
@@ -5,5 +5,5 @@ package com.mashup.zuzu.data.request
  */
 data class UpdateUsersRequest(
     val nickname: String,
-    val profile_id: Int
+    val profile_id: Long
 )
